refactor: split share.service.ts into invitation/management sub-services
ShareService (836 lines) was the largest remaining backend service with
no split precedent, unlike files/conversations/deployments/toolsets/prompts.
Splits it into ShareInvitationService (createShareLink, acceptInvitation)
and ShareManagementService (discardShared, getRecipientsCount,
revokeShared), plus a shared utils/share-resource.util.ts and a thin
ShareService facade. No REST contract or behavior changes.
