+++ apps/chat-api/src/share/share.service.ts
@@ -1,836 +1,75 @@
-import type { components } from '@epam/ai-dial-typescript-sdk';
-import {
-  BadGatewayException,
-  ForbiddenException,
-  Injectable,
-  Logger,
-  NotFoundException,
-} from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import {
-  handleDialFetchError,
-  handleDialSdkError,
-  mapDialHttpStatus,
-} from '../common/dial/dial-error.mapper';
-import { getBearerAuthHeaders } from '../common/utils/auth-header';
-import { encodeDialResourcePath } from '../common/utils/encode-dial-path';
-import {
-  countRecipientsByUrl,
-  resolveRecipientsCount,
-} from '../common/utils/resource-ownership';
-import { safeDecodeURIComponent } from '../common/utils/uri';
-import { EnvironmentVariables } from '../config/environment.config';
-import { PUBLIC_BUCKET } from '../conversations/constants/conversation.constants';
-import { resolveConversationLocation } from '../conversations/utils/conversation.utils';
-import { DeploymentsService } from '../deployments/deployments.service';
-import { DialClientService } from '../dial/dial-client.service';
-import { isPromptResourceUrl } from '../prompts/utils/prompt-mapper.util';
-import { getResourceBucket } from '../publish/publish-target.util';
-import { SkillsLookupService } from '../skills/lookup/skills-lookup.service';
-import { ToolsetsService } from '../toolsets/toolsets.service';
+import { Injectable } from '@nestjs/common';
 import { AcceptInvitationResponseDto } from './dto/accept-invitation-response.dto';
-import { CreateShareLinkDto, ShareAccess } from './dto/create-share-link.dto';
+import { CreateShareLinkDto } from './dto/create-share-link.dto';
 import { DiscardSharedCatalogItemResponseDto } from './dto/discard-shared-catalog-item.dto';
 import { RevokeSharedAccessResponseDto } from './dto/revoke-shared-access.dto';
 import { ShareLinkResponseDto } from './dto/share-link-response.dto';
 import { ShareRecipientsResponseDto } from './dto/share-recipients.dto';
+import { ShareInvitationService } from './invitation/share-invitation.service';
+import { ShareManagementService } from './management/share-management.service';
 
-type ResourceAccessType = components['schemas']['ResourceAccessType'];
-type ResourceKind = components['schemas']['ResourceTypes'];
-
-/*
- * `discardShared`/`revokeShared`/`getRecipientsCount` always receive an
- * `itemId` that is already a full DIAL Core resource url — the DTO's regex
- * validation enforces one of these prefixes before this ever runs — so
- * `resolveResourceKind` below always finds a match. This table exists purely
- * to translate that prefix into the `resourceTypes` filter
- * `getSharedResources` expects.
- */
-const RESOURCE_KIND_BY_PREFIX: [prefix: string, kind: ResourceKind][] = [
-  ['applications/', 'APPLICATION'],
-  ['toolsets/', 'TOOL_SET'],
-  ['conversations/', 'CONVERSATION'],
-  ['skills/', 'SKILL'],
-  ['prompts/', 'PROMPT'],
-];
-
-const resolveResourceKind = (itemId: string): ResourceKind => {
-  const match = RESOURCE_KIND_BY_PREFIX.find(([prefix]) =>
-    itemId.startsWith(prefix),
-  );
-  if (!match)
-    throw new Error(`Unrecognized resource kind for itemId: ${itemId}`);
-  return match[1];
-};
-
-/*
- * Prompts are the one resource kind whose public `itemId` is deliberately
- * decoded back to a raw, human-readable DIAL Core resource path
- * (`buildPromptId` rebuilds it from a `safeDecodeURIComponent`d metadata
- * url — see `urlToPromptPath` — so folder/prompt names with spaces stay
- * literal wherever the id is displayed). Every other kind's listing passes
- * DIAL Core's own metadata url straight through unchanged (conversations:
- * `conversation-listing.service.ts`; applications/toolsets: `raw.id` in
- * their mapper utils), so their `itemId` already matches whatever encoded
- * or unencoded form DIAL Core itself uses — re-encoding it here would risk
- * disagreeing with that native form. Only prompts need re-encoding back to
- * DIAL Core's canonical form before this direct SDK call boundary.
- *
- * TODO: once prompts stop decoding at listing time (aligning them with the
- * conversations pattern — using metadata's native `name`/`parentPath`
- * fields instead of deriving from a decoded path), this conditional goes
- * away entirely and `itemId` can be used as-is for every kind.
- */
-const toShareResourceUrl = (itemId: string): string =>
-  isPromptResourceUrl(itemId) ? encodeDialResourcePath(itemId) : itemId;
-
-/*
- * DIAL Core's `shareResource` endpoint does not return an expiry; the link
- * expiry is a fixed platform default rather than something DIAL Core reports
- * back per-request.
- */
-const SHARE_LINK_EXPIRES_IN_DAYS = 3;
-
-const ACCESS_PERMISSIONS: Record<ShareAccess, ResourceAccessType[]> = {
-  [ShareAccess.View]: ['READ'],
-  [ShareAccess.Edit]: ['READ', 'WRITE'],
-};
-
-/*
- * The generated share link must point at a frontend route the SPA can
- * render (which then accepts the invitation and redirects into the shared
- * resource), not at DIAL Core's own `/v1/invitations/{id}` API path. Which
- * frontend route depends on the resource kind: catalog entities land on the
- * catalog accept-invitation route, conversations on the conversation one, so
- * each redirects into the right place after acceptance.
- *
- * These string literals MUST stay in sync with `ROUTES.SharedInvitation` and
- * `ROUTES.ConversationSharedInvitation` in `apps/chat/src/types/routes.ts` —
- * apps cannot import each other's route constants across the Nx module
- * boundary, so there is no single shared source of truth. Renaming either
- * frontend route without updating the matching constant here (or vice versa)
- * compiles cleanly but silently breaks the generated share link. The tests
- * in `share.service.spec.ts` assert the exact URL prefix for both kinds —
- * update them alongside any rename here.
- */
-const CATALOG_SHARE_INVITATION_ROUTE_PATH = '/catalog/shared';
-const CONVERSATION_SHARE_INVITATION_ROUTE_PATH = '/conversations/shared';
-
-/** DIAL Core conversation resource paths always start with this prefix. */
-const CONVERSATION_RESOURCE_PREFIX = 'conversations/';
-const FILE_RESOURCE_PREFIX = 'files/';
-
-interface AnnotationWithAttachment {
-  body?: {
-    source?: {
-      attachment?: unknown;
-    };
-  };
-}
-
-const isRecord = (value: unknown): value is Record<string, unknown> =>
-  typeof value === 'object' && value !== null;
-
-const collectAttachmentResourceUrls = (
-  attachments: unknown,
-  resourceUrls: Set<string>,
-): void => {
-  if (!Array.isArray(attachments)) return;
-
-  for (const attachment of attachments) {
-    if (!isRecord(attachment)) continue;
-
-    for (const field of ['url', 'reference_url'] as const) {
-      const url = attachment[field];
-      if (typeof url !== 'string') continue;
-
-      const resourceUrl = url.split('#', 1)[0];
-      if (resourceUrl.startsWith(FILE_RESOURCE_PREFIX)) {
-        resourceUrls.add(resourceUrl);
-      }
-    }
-  }
-};
-
-/** Collects unique DIAL file resources referenced by messages, stages, and citations. */
-const collectConversationResourceUrls = (conversation: unknown): string[] => {
-  if (!isRecord(conversation) || !Array.isArray(conversation.messages)) {
-    return [];
-  }
-
-  const resourceUrls = new Set<string>();
-  for (const message of conversation.messages) {
-    if (!isRecord(message) || !isRecord(message.custom_content)) continue;
-
-    collectAttachmentResourceUrls(
-      message.custom_content.attachments,
-      resourceUrls,
-    );
-
-    const stages = message.custom_content.stages;
-    if (Array.isArray(stages)) {
-      for (const stage of stages) {
-        if (isRecord(stage)) {
-          collectAttachmentResourceUrls(stage.attachments, resourceUrls);
-        }
-      }
-    }
-
-    const annotations = message.custom_content.annotations;
-    if (!Array.isArray(annotations)) continue;
-    for (const annotation of annotations) {
-      const attachment = (annotation as AnnotationWithAttachment)?.body?.source
-        ?.attachment;
-      collectAttachmentResourceUrls([attachment], resourceUrls);
-    }
-  }
-
-  return [...resourceUrls];
-};
-
-const getInvitationRoutePath = (itemId: string): string =>
-  itemId.startsWith(CONVERSATION_RESOURCE_PREFIX)
-    ? CONVERSATION_SHARE_INVITATION_ROUTE_PATH
-    : CATALOG_SHARE_INVITATION_ROUTE_PATH;
-
-/*
- * DIAL Core rejects `?accept=true` with 400 and a body like
- * `"Resource <id> already belong to you"` when the invited user already owns
- * (or previously accepted) the shared resource — e.g. opening your own share
- * link, or reopening a link you already accepted. That's not a real failure:
- * the user already has access, so the accept call is a no-op and the UI
- * should still open the resource's details panel rather than showing an
- * error. Detected by substring match since DIAL Core doesn't give this case
- * its own status code or a machine-readable error field.
- */
-const isAlreadyOwnedError = (errorBody: unknown): boolean =>
-  typeof errorBody === 'string' &&
-  errorBody.toLowerCase().includes('already belong');
-
-/** Creates share links for DIAL Core resources (catalog entities or conversations) by proxying DIAL Core's resource-sharing API. */
+/** Facade over share-invitation and share-management operations for DIAL Core resources (catalog entities, prompts, or conversations). */
 @Injectable()
 export class ShareService {
-  private readonly logger = new Logger(ShareService.name);
-  private readonly appOrigin: string;
-
   constructor(
-    private readonly dialClient: DialClientService,
-    private readonly configService: ConfigService<EnvironmentVariables>,
-    private readonly deploymentsService: DeploymentsService,
-    private readonly toolsetsService: ToolsetsService,
-    private readonly skillsLookupService: SkillsLookupService,
-  ) {
-    const callbackBaseUrl = this.configService.get('AUTH_CALLBACK_BASE_URL', {
-      infer: true,
-    });
-    this.appOrigin = new URL(callbackBaseUrl).origin;
-  }
+    private readonly shareInvitationService: ShareInvitationService,
+    private readonly shareManagementService: ShareManagementService,
+  ) {}
 
-  /*
-   * DIAL Core's `invitationLink` (e.g. `/v1/invitations/{id}`) is an API
-   * path, not a page the SPA can render, and it's host-relative to DIAL
-   * Core rather than the frontend's public origin. Only the trailing id
-   * segment is reused, to build an absolute frontend URL that lands on the
-   * SPA's own accept-invitation route.
-   */
-  private buildInvitationUrl(invitationLink: string, itemId: string): string {
-    const { pathname } = new URL(invitationLink, this.appOrigin);
-    const invitationId = pathname.split('/').filter(Boolean).pop();
-    if (!invitationId) {
-      throw new BadGatewayException(
-        'DIAL Core returned an invalid invitation link',
-      );
-    }
-    return `${this.appOrigin}${getInvitationRoutePath(itemId)}/${invitationId}`;
-  }
-
-  private async getRelatedResourceUrls(
-    accessToken: string,
-    sessionBucket: string,
-    resourceUrl: string,
-  ): Promise<string[]> {
-    if (!resourceUrl.startsWith(CONVERSATION_RESOURCE_PREFIX)) return [];
-
-    const conversationPath = resourceUrl.slice(
-      CONVERSATION_RESOURCE_PREFIX.length,
-    );
-    const { bucket, subPath } = resolveConversationLocation(
-      conversationPath,
-      sessionBucket,
-    );
-
-    let result;
-    try {
-      result = await this.dialClient.client.getConversation(
-        bucket,
-        encodeDialResourcePath(subPath),
-        { headers: getBearerAuthHeaders(accessToken) },
-      );
-    } catch (err) {
-      return handleDialSdkError(
-        err,
-        'load conversation resources before sharing',
-        this.logger,
-      );
-    }
-
-    if (result.error != null) {
-      return handleDialSdkError(
-        result.error,
-        'load conversation resources before sharing',
-        this.logger,
-        result.response,
-      );
-    }
-    if (result.data == null) {
-      this.logger.error(
-        `DIAL Core returned an empty conversation for resourceUrl=${resourceUrl}`,
-      );
-      throw new BadGatewayException('DIAL Core returned an empty conversation');
-    }
-
-    /* See openspec/specs/conversation-share/spec.md — "Related file resources outside the conversation's own bucket are dropped". */
-    return collectConversationResourceUrls(result.data).filter((url) => {
-      const fileBucket = getResourceBucket(url);
-      return fileBucket === bucket || fileBucket === PUBLIC_BUCKET;
-    });
-  }
-
-  /**
-   * Creates a share link for a DIAL Core resource (catalog entity, prompt, or conversation).
-   *
-   * @throws {BadGatewayException} When DIAL Core returns an error response
-   * @throws {ServiceUnavailableException} When DIAL Core is unreachable or times out
-   */
-  async createShareLink(
+  createShareLink(
     accessToken: string,
     bucket: string,
-    { itemId, access }: CreateShareLinkDto,
+    dto: CreateShareLinkDto,
   ): Promise<ShareLinkResponseDto> {
-    const resourceUrl = toShareResourceUrl(itemId);
-    const permissions = Array.from(
-      new Set(access.flatMap((level) => ACCESS_PERMISSIONS[level])),
-    );
-    const relatedResourceUrls = await this.getRelatedResourceUrls(
+    return this.shareInvitationService.createShareLink(
       accessToken,
       bucket,
-      resourceUrl,
+      dto,
     );
-    const requestBody = {
-      invitationType: 'LINK' as const,
-      resources: [resourceUrl, ...relatedResourceUrls].map((url) => ({
-        url,
-        permissions,
-      })),
-    };
-    this.logger.debug(
-      `Requesting share link from DIAL Core: ${JSON.stringify(requestBody)}`,
-    );
-
-    let result;
-    try {
-      result = await this.dialClient.client.shareResource({
-        headers: getBearerAuthHeaders(accessToken),
-        body: requestBody,
-      });
-    } catch (err) {
-      return handleDialFetchError(err, 'create share link', this.logger, 0);
-    }
-
-    if (result.error) {
-      return mapDialHttpStatus(
-        result.response.status,
-        'create share link',
-        this.logger,
-        result.error,
-      );
-    }
-
-    this.logger.debug(
-      `DIAL Core share link response for itemId=${resourceUrl}: ${JSON.stringify(result.data)}`,
-    );
-
-    const invitationLink = result.data?.invitationLink;
-    if (invitationLink == null) {
-      this.logger.error(
-        `DIAL Core returned an empty invitation link for itemId=${resourceUrl}`,
-      );
-      throw new BadGatewayException(
-        'DIAL Core returned an empty invitation link',
-      );
-    }
-
-    this.logger.debug(`Created share link for itemId=${resourceUrl}`);
-
-    return {
-      url: this.buildInvitationUrl(invitationLink, resourceUrl),
-      expiresInDays: SHARE_LINK_EXPIRES_IN_DAYS,
-      access,
-    };
   }
 
-  /**
-   * Accepts a share invitation via DIAL Core, granting the authenticated
-   * user its access level, and returns the shared entity's identifier.
-   *
-   * @throws {BadGatewayException} When DIAL Core returns an error response
-   * @throws {ServiceUnavailableException} When DIAL Core is unreachable or times out
-   */
-  async acceptInvitation(
+  acceptInvitation(
     accessToken: string,
     invitationId: string,
     userSub: string,
     bucket: string,
   ): Promise<AcceptInvitationResponseDto> {
-    this.logger.debug(
-      `Peeking invitation from DIAL Core for invitationId=${invitationId}`,
-    );
-
-    /*
-     * DIAL Core's `?accept=true` performs the accept side effect but returns
-     * an empty 200 body rather than the `Invitation` payload its OpenAPI
-     * schema declares — confirmed by logging the raw response body, which
-     * came back empty. So the shared resource's itemId has to be read from a
-     * separate, non-accepting peek call before the accepting call runs.
-     */
-    let peekResult;
-    try {
-      peekResult = await this.dialClient.client.getInvitation(invitationId, {
-        headers: getBearerAuthHeaders(accessToken),
-      });
-    } catch (err) {
-      return handleDialFetchError(err, 'peek invitation', this.logger, 0);
-    }
-
-    if (peekResult.error) {
-      return mapDialHttpStatus(
-        peekResult.response.status,
-        'peek invitation',
-        this.logger,
-        peekResult.error,
-      );
-    }
-
-    const resources = peekResult.data?.resources ?? [];
-    const itemId =
-      resources.find((r) => !r.url?.startsWith(FILE_RESOURCE_PREFIX))?.url ??
-      resources[0]?.url;
-    if (itemId == null) {
-      this.logger.error(
-        `DIAL Core returned an invitation with no shared resource for invitationId=${invitationId}`,
-      );
-      throw new BadGatewayException(
-        'DIAL Core returned an invitation with no shared resource',
-      );
-    }
-
-    this.logger.debug(
-      `Accepting invitation from DIAL Core for invitationId=${invitationId}, itemId=${itemId}`,
-    );
-
-    let acceptResult;
-    try {
-      acceptResult = await this.dialClient.client.getInvitation(invitationId, {
-        headers: getBearerAuthHeaders(accessToken),
-        params: { query: { accept: true } },
-      });
-    } catch (err) {
-      return handleDialFetchError(err, 'accept invitation', this.logger, 0);
-    }
-
-    if (
-      acceptResult.error &&
-      !(
-        acceptResult.response.status === 400 &&
-        isAlreadyOwnedError(acceptResult.error)
-      )
-    ) {
-      return mapDialHttpStatus(
-        acceptResult.response.status,
-        'accept invitation',
-        this.logger,
-        acceptResult.error,
-      );
-    }
-
-    this.logger.debug(
-      acceptResult.error
-        ? `Invitation invitationId=${invitationId} resolves to a resource the user already owns; treating as accepted`
-        : `Accepted invitation for invitationId=${invitationId}`,
-    );
-
-    /*
-     * The deployments/toolsets lists are cached per user for 30s
-     * (`DeploymentsService`/`ToolsetsService`). Without invalidating them
-     * here, the frontend's post-accept refetch can still serve the
-     * pre-share snapshot, so the catalog's details panel silently fails to
-     * find the newly shared item.
-     */
-    await Promise.all([
-      this.deploymentsService.invalidateListCache(userSub),
-      this.toolsetsService.invalidateListCache(userSub),
-    ]);
-
-    const summary = await this.resolveSharedItemSummary(
-      itemId,
+    return this.shareInvitationService.acceptInvitation(
       accessToken,
+      invitationId,
       userSub,
       bucket,
     );
-
-    return { itemId, ...summary };
   }
 
-  /**
-   * Resolves the just-accepted item's list-item summary by id, so the
-   * frontend can show it immediately instead of depending on a subsequent
-   * bulk deployments/toolsets list refresh reflecting the grant — DIAL Core
-   * does not guarantee that a `listDeployments`/`getSharedResources` call
-   * immediately after an accept already includes the newly shared resource.
-   *
-   * Best-effort: any failure here (resolution error, or DIAL Core genuinely
-   * has no match yet) resolves to an empty object rather than throwing — the
-   * invitation was already successfully accepted upstream, so a summary
-   * lookup failure must not fail the whole `acceptInvitation` call.
-   */
-  private async resolveSharedItemSummary(
-    itemId: string,
-    accessToken: string,
-    userSub: string,
-    bucket: string,
-  ): Promise<
-    Pick<
-      AcceptInvitationResponseDto,
-      'sharedDeployment' | 'sharedToolset' | 'sharedSkill'
-    >
-  > {
-    try {
-      if (itemId.startsWith('skills/')) {
-        const sharedSkill = await this.skillsLookupService.resolveSkillItem(
-          itemId,
-          accessToken,
-          bucket,
-        );
-        return sharedSkill ? { sharedSkill } : {};
-      }
-
-      /*
-       * A prompt has no deployments/toolsets list entry to summarise — the
-       * frontend picks it up from its own prompts refetch instead.
-       */
-      if (isPromptResourceUrl(itemId)) return {};
-
-      if (itemId.startsWith('toolsets/')) {
-        const sharedToolset = await this.toolsetsService.resolveToolsetItem(
-          userSub,
-          accessToken,
-          itemId,
-        );
-        return sharedToolset ? { sharedToolset } : {};
-      }
-
-      const sharedDeployment =
-        await this.deploymentsService.resolveDeploymentItem(
-          itemId,
-          accessToken,
-          bucket,
-        );
-      if (sharedDeployment) return { sharedDeployment };
-
-      if (!itemId.startsWith('applications/')) {
-        const sharedToolset = await this.toolsetsService.resolveToolsetItem(
-          userSub,
-          accessToken,
-          itemId,
-        );
-        if (sharedToolset) return { sharedToolset };
-      }
-
-      return {};
-    } catch (err) {
-      this.logger.warn(
-        `Failed to resolve shared item summary for itemId=${itemId}`,
-        err,
-      );
-      return {};
-    }
-  }
-
-  /*
-   * DIAL Core's `discardSharedResources` treats a resource that was never
-   * shared with the caller as an idempotent no-op — it answers `200` rather
-   * than `403` — so a blind pass-through of its response would silently
-   * report success without actually removing (or ever having granted) any
-   * access. `discardShared` below cross-checks this against
-   * `getSharedResources` (the same call the deployments/toolsets list
-   * endpoints use to derive `sharedWithMe`), taken *before* calling discard,
-   * to tell a genuine discard apart from that silent no-op.
-   */
-  private async isSharedWithCaller(
-    resourceUrl: string,
-    accessToken: string,
-  ): Promise<boolean> {
-    let result;
-    try {
-      result = await this.dialClient.client.getSharedResources({
-        headers: getBearerAuthHeaders(accessToken),
-        body: {
-          resourceTypes: [resolveResourceKind(resourceUrl)],
-          with: 'me',
-        },
-      });
-    } catch (err) {
-      return handleDialFetchError(
-        err,
-        'share.discardShared (verify shared)',
-        this.logger,
-        0,
-      );
-    }
-
-    if (result.error) {
-      return mapDialHttpStatus(
-        result.response.status,
-        'share.discardShared (verify shared)',
-        this.logger,
-        result.error,
-      );
-    }
-
-    /*
-     * Both encodings are tried, same as `getRecipientsCount`: list ids and
-     * DIAL Core share urls differ in percent-encoding for some resource
-     * types (conversations in particular), so an exact match against only
-     * one form would wrongly report "not shared" for a genuinely-shared
-     * resource whose name needs percent-encoding (a space, for instance).
-     */
-    const decodedResourceUrl = safeDecodeURIComponent(resourceUrl);
-    return (result.data?.resources ?? []).some(
-      (resource) =>
-        resource.url === resourceUrl || resource.url === decodedResourceUrl,
-    );
-  }
-
-  /**
-   * Discards the calling user's own access to a catalog resource shared with
-   * them, via DIAL Core `discardSharedResources`. This only affects the
-   * caller — removing access for everyone else is the owner-side
-   * {@link ShareService.revokeShared} operation.
-   *
-   * `itemId` is always a full DIAL Core resource path, including for a
-   * shared-with-the-caller prompt — it already embeds its owner's bucket,
-   * not the caller's own.
-   *
-   * @throws {ForbiddenException} When the resource is not shared with the caller
-   * @throws {NotFoundException} When the resource does not exist
-   * @throws {BadGatewayException} When DIAL Core returns an error response
-   * @throws {ServiceUnavailableException} When DIAL Core is unreachable or times out
-   */
-  async discardShared(
+  discardShared(
     itemId: string,
     accessToken: string,
     userSub: string,
   ): Promise<DiscardSharedCatalogItemResponseDto> {
-    this.logger.log('Discard shared resource started');
-
-    const resourceUrl = toShareResourceUrl(itemId);
-
-    /*
-     * Read before calling discard, not after: once discard runs, a resource
-     * that really was shared is no longer shared either, so checking
-     * afterwards can't tell a genuine discard apart from the no-op case this
-     * is meant to catch.
-     */
-    const wasSharedWithCaller = await this.isSharedWithCaller(
-      resourceUrl,
+    return this.shareManagementService.discardShared(
+      itemId,
       accessToken,
+      userSub,
     );
-
-    let result;
-    try {
-      result = await this.dialClient.client.discardSharedResources({
-        headers: getBearerAuthHeaders(accessToken),
-        body: { resources: [{ url: resourceUrl }] },
-      });
-    } catch (err) {
-      return handleDialFetchError(err, 'share.discardShared', this.logger, 0);
-    }
-
-    if (result.error) {
-      /*
-       * DIAL Core has no dedicated status for "itemId is well-formed but
-       * doesn't resolve to any resource" — it answers a generic `400` here,
-       * indistinguishable at the wire level from a truly malformed request.
-       * `DiscardSharedCatalogItemDto` already rejects malformed itemIds
-       * before this method runs, so any `400` reaching this point can only
-       * be DIAL Core's not-found case; map it to the `404` the
-       * `catalog-unshare` spec requires.
-       */
-      if (result.response.status === 400) {
-        throw new NotFoundException('Resource does not exist');
-      }
-      return mapDialHttpStatus(
-        result.response.status,
-        'share.discardShared',
-        this.logger,
-        result.error,
-      );
-    }
-
-    if (!wasSharedWithCaller) {
-      this.logger.warn(
-        `Discard shared resource rejected: itemId=${itemId} is not shared with the caller`,
-      );
-      throw new ForbiddenException('Resource is not shared with the caller');
-    }
-
-    await Promise.all([
-      this.deploymentsService.invalidateListCache(userSub),
-      this.toolsetsService.invalidateListCache(userSub),
-    ]);
-
-    this.logger.log('Discard shared resource completed: success=true');
-    return { success: true };
   }
 
-  /**
-   * Counts how many users currently hold shared access to one resource the
-   * caller owns, via DIAL Core `getSharedResources({ with: 'others' })`.
-   *
-   * Answered per resource, on demand — the frontend calls this when the owner
-   * opens the action menu that offers {@link ShareService.revokeShared}, so a
-   * count is never older than the menu showing it. DIAL Core has no
-   * single-resource variant of this query, so the whole `with: 'others'` set
-   * for the resource's kind is fetched and the one entry is picked out of it.
-   *
-   * `sharedWith` only lists users who *accepted* an invitation, so `0` means
-   * "nobody holds access", not "never shared" — an issued but unopened share
-   * link contributes nothing.
-   *
-   * `itemId` is always a full DIAL Core resource path, and only its owner
-   * ever sees its recipient count.
-   *
-   * @throws {BadGatewayException} When DIAL Core returns an error response
-   * @throws {ServiceUnavailableException} When DIAL Core is unreachable or times out
-   */
-  async getRecipientsCount(
+  getRecipientsCount(
     itemId: string,
     accessToken: string,
   ): Promise<ShareRecipientsResponseDto> {
-    const resourceUrl = toShareResourceUrl(itemId);
-
-    let result;
-    try {
-      result = await this.dialClient.client.getSharedResources({
-        headers: getBearerAuthHeaders(accessToken),
-        body: {
-          resourceTypes: [resolveResourceKind(resourceUrl)],
-          with: 'others',
-          includeUserInfo: true,
-        },
-      });
-    } catch (err) {
-      return handleDialFetchError(
-        err,
-        'share.getRecipientsCount',
-        this.logger,
-        0,
-      );
-    }
-
-    if (result.error) {
-      return mapDialHttpStatus(
-        result.response.status,
-        'share.getRecipientsCount',
-        this.logger,
-        result.error,
-      );
-    }
-
-    const counts = countRecipientsByUrl(
-      (result.data?.resources ?? []) as {
-        url?: string;
-        sharedWith?: unknown[];
-      }[],
-    );
-    /*
-     * Both encodings are tried: list ids and DIAL Core share urls differ in
-     * percent-encoding for some resource types (conversations in particular).
-     */
-    const recipientsCount =
-      resolveRecipientsCount(
-        counts,
-        resourceUrl,
-        safeDecodeURIComponent(resourceUrl),
-      ) ?? 0;
-
-    return { itemId, recipientsCount };
+    return this.shareManagementService.getRecipientsCount(itemId, accessToken);
   }
 
-  /**
-   * Revokes every outstanding share grant on a resource the caller owns, via
-   * DIAL Core `revokeSharedResources`. This affects all recipients at once —
-   * DIAL Core's request carries no subject field, so a single recipient
-   * cannot be targeted. Discarding only the caller's own access is the
-   * recipient-side {@link ShareService.discardShared} operation.
-   *
-   * Unlike `discardShared` there is no pre-flight `getSharedResources` check:
-   * for an owner, a resource that currently has no recipients already has the
-   * requested outcome, so DIAL Core's no-op success is the correct answer
-   * rather than something to surface as an error. Ownership itself is
-   * enforced by DIAL Core, which answers `403` for a non-owner.
-   *
-   * `itemId` is always a full DIAL Core resource path, and only its owner
-   * can revoke its shared access.
-   *
-   * @throws {NotFoundException} When the resource does not exist
-   * @throws {BadGatewayException} When DIAL Core returns an error response
-   * @throws {ServiceUnavailableException} When DIAL Core is unreachable or times out
-   */
-  async revokeShared(
+  revokeShared(
     itemId: string,
     accessToken: string,
     userSub: string,
   ): Promise<RevokeSharedAccessResponseDto> {
-    this.logger.log('Revoke shared access started');
-
-    const resourceUrl = toShareResourceUrl(itemId);
-
-    let result;
-    try {
-      result = await this.dialClient.client.revokeSharedResources({
-        headers: getBearerAuthHeaders(accessToken),
-        body: { resources: [{ url: resourceUrl }] },
-      });
-    } catch (err) {
-      return handleDialFetchError(err, 'share.revokeShared', this.logger, 0);
-    }
-
-    if (result.error) {
-      /*
-       * Same reasoning as `discardShared`: DIAL Core has no dedicated status
-       * for "itemId is well-formed but resolves to no resource" and answers a
-       * generic `400`. `RevokeSharedAccessDto` already rejects malformed
-       * itemIds before this method runs, so any `400` reaching here can only
-       * be the not-found case.
-       */
-      if (result.response.status === 400) {
-        throw new NotFoundException('Resource does not exist');
-      }
-      return mapDialHttpStatus(
-        result.response.status,
-        'share.revokeShared',
-        this.logger,
-        result.error,
-      );
-    }
-
-    await Promise.all([
-      this.deploymentsService.invalidateListCache(userSub),
-      this.toolsetsService.invalidateListCache(userSub),
-    ]);
-
-    this.logger.log('Revoke shared access completed: success=true');
-    return { success: true };
+    return this.shareManagementService.revokeShared(
+      itemId,
+      accessToken,
+      userSub,
+    );
   }
 }
