+++ openspec/changes/archive/2026-09-05-split-share-service/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-09-05
