+++ openspec/specs/catalog-unshare/spec.md
@@ -78,13 +78,13 @@ The endpoint SHALL:
 - Require a valid session; respond `401 Unauthorized` when no session is present.
 - Accept `DiscardSharedCatalogItemDto { itemId: string }` validated via NestJS `ValidationPipe` (whitelist, forbidNonWhitelisted, transform); `itemId` SHALL be a non-empty string, max length 2048, validated with the existing `IsValidFilePath` validator and an `@Matches` allowlist restricted to `applications/{bucket}/{path}`, `toolsets/{bucket}/{path}`, `conversations/{bucket}/{path}`, `skills/{bucket}/{path}`, **or `prompts/{bucket}/{path}`**. Other DIAL resource types and incomplete paths SHALL be rejected before calling DIAL Core. Because the shared `@Matches` pattern's trailing segment is unrestricted (it also matches deeper nested paths), a `skills/`-prefixed `itemId` containing a `/files/` segment SHALL be additionally rejected by a supplementary validator on `DiscardSharedCatalogItemDto`, so only whole-skill URLs (never a single in-skill file URL) are accepted — skills remain whole-resource units for sharing. `DiscardSharedCatalogItemDto` carries no `resourceKind` or `bucket` field — every `itemId`, prompts included, is a self-sufficient full resource path, and no per-resource-type qualification step exists on this DTO.
 - Use the session `accessToken` as the Bearer credential when calling DIAL Core.
-- Call SDK `discardSharedResources({ headers, body: { resources: [{ url: itemId }] } })` with no bucket/path reconstruction — `itemId` is passed through unmodified as the resource `url`, matching the existing `createShareLink` pattern (`share.service.ts`).
+- Call SDK `discardSharedResources({ headers, body: { resources: [{ url: itemId }] } })` with no bucket/path reconstruction — `itemId` is passed through unmodified as the resource `url`, matching the existing `createShareLink` pattern (`apps/chat-api/src/share/invitation/share-invitation.service.ts`).
 - Rely on DIAL Core to enforce that the resource is currently shared with the caller; a resource not shared with the caller SHALL surface as `403 Forbidden` via `mapDialHttpStatus`, not a silent 200.
-- Resolve the DIAL Core `resourceTypes` filter used by the pre-discard "was this shared with me" check (`ShareService.isSharedWithCaller`) via `RESOURCE_KIND_BY_PREFIX` (`share.service.ts`), which SHALL include a `['skills/', 'SKILL']` entry and a `['prompts/', 'PROMPT']` entry alongside the existing `applications/` → `APPLICATION`, `toolsets/` → `TOOL_SET`, and `conversations/` → `CONVERSATION` entries.
-- On success, invalidate both `DeploymentsService.invalidateListCache(userSub)` and `ToolsetsService.invalidateListCache(userSub)` before responding, mirroring the existing invalidation call in `ShareService.acceptInvitation`. This invalidation runs unconditionally regardless of `itemId` type; conversations, skills, and prompts have no equivalent server-side list cache today, so for those `itemId` types this invalidation is a harmless no-op.
+- Resolve the DIAL Core `resourceTypes` filter used by the pre-discard "was this shared with me" check (`ShareManagementService.isSharedWithCaller`, `apps/chat-api/src/share/management/share-management.service.ts`) via `RESOURCE_KIND_BY_PREFIX` (`apps/chat-api/src/share/utils/share-resource.util.ts`), which SHALL include a `['skills/', 'SKILL']` entry and a `['prompts/', 'PROMPT']` entry alongside the existing `applications/` → `APPLICATION`, `toolsets/` → `TOOL_SET`, and `conversations/` → `CONVERSATION` entries.
+- On success, invalidate both `DeploymentsService.invalidateListCache(userSub)` and `ToolsetsService.invalidateListCache(userSub)` before responding, mirroring the existing invalidation call in `ShareInvitationService.acceptInvitation`. This invalidation runs unconditionally regardless of `itemId` type; conversations, skills, and prompts have no equivalent server-side list cache today, so for those `itemId` types this invalidation is a harmless no-op.
 - Respond `200 OK` with `DiscardSharedCatalogItemResponseDto { success: true }` on success.
 - Apply `@Throttle({ default: { limit: 10, ttl: 60000 } })`, matching the file-manager `discard-shared` endpoint's stricter-than-share-creation posture.
-- Map upstream failures via the fetch-shaped `mapDialHttpStatus`/`handleDialFetchError` pair (consistent with `ShareService`'s other methods): DIAL Core 400 → 400, 401 → 401, 403 → 403, 404 → 404, 429 → 429, 5xx → 502, network/timeout → 503.
+- Map upstream failures via the fetch-shaped `mapDialHttpStatus`/`handleDialFetchError` pair (consistent with the share domain's other methods): DIAL Core 400 → 400, 401 → 401, 403 → 403, 404 → 404, 429 → 429, 5xx → 502, network/timeout → 503.
 - Not cache the mutation response itself.
 - Log structured success/failure messages (e.g. `Discard shared resource started`, `Discard shared resource completed: success=true`, `DIAL Core returned <status> for share.discardShared`) without the access token, invitation links, full resource path, or any other user data beyond a safe operation identifier.
