+++ openspec/specs/share-revoke-access/spec.md
@@ -18,14 +18,14 @@ The endpoint SHALL:
 - Call SDK `revokeSharedResources({ headers, body: { resources: [{ url: itemId }] } })`, passing `itemId` through unmodified with no bucket/path reconstruction.
 - NOT perform any pre-flight `getSharedResources` check. Unlike `discardShared`, a resource with no current recipients is a legitimate no-op success for the owner, not a condition to surface as an error.
 - Rely on DIAL Core to enforce ownership; a caller who does not own the resource SHALL surface as `403 Forbidden` via `mapDialHttpStatus`.
-- On success, invalidate both `DeploymentsService.invalidateListCache(userSub)` and `ToolsetsService.invalidateListCache(userSub)` before responding, unconditionally regardless of `itemId` type, mirroring `ShareService.discardShared`. Conversations, skills, and prompts have no equivalent server-side list cache, so for those `itemId` types this is a harmless no-op.
+- On success, invalidate both `DeploymentsService.invalidateListCache(userSub)` and `ToolsetsService.invalidateListCache(userSub)` before responding, unconditionally regardless of `itemId` type, mirroring `ShareManagementService.discardShared`. Conversations, skills, and prompts have no equivalent server-side list cache, so for those `itemId` types this is a harmless no-op.
 - Respond `200 OK` with `RevokeSharedAccessResponseDto { success: true }`. DIAL Core returns an empty 200 body for this operation, so the response is synthesized by the BFF.
 - Apply `@Throttle({ default: { limit: 10, ttl: 60000 } })`, matching the discard endpoint's posture.
 - Map upstream failures via the fetch-shaped `mapDialHttpStatus` / `handleDialFetchError` pair: DIAL Core 400 → 404 (`'Resource does not exist'`, since the DTO already rejects malformed itemIds so a Core 400 can only mean an unresolvable resource — same reasoning as `discardShared`), 401 → 401, 403 → 403, 404 → 404, 429 → 429, 5xx → 502, network/timeout → 503.
 - Not cache the mutation response.
 - Log structured start/completion messages (e.g. `Revoke shared access started`, `Revoke shared access completed: success=true`) without the access token, invitation links, full resource path, or any other user data.
 
-Resolve the DIAL Core `resourceTypes` filter for `revokeShared` via the existing `RESOURCE_KIND_BY_PREFIX` map (`share.service.ts`), which already includes a `['skills/', 'SKILL']` entry and a `['prompts/', 'PROMPT']` entry alongside `applications/` → `APPLICATION`, `toolsets/` → `TOOL_SET`, and `conversations/` → `CONVERSATION` — no change to this map is required by this capability, only to the DTO allowlist that gates whether a `prompts/` `itemId` reaches it.
+Resolve the DIAL Core `resourceTypes` filter for `revokeShared` via the existing `RESOURCE_KIND_BY_PREFIX` map (`apps/chat-api/src/share/utils/share-resource.util.ts`), which already includes a `['skills/', 'SKILL']` entry and a `['prompts/', 'PROMPT']` entry alongside `applications/` → `APPLICATION`, `toolsets/` → `TOOL_SET`, and `conversations/` → `CONVERSATION` — no change to this map is required by this capability, only to the DTO allowlist that gates whether a `prompts/` `itemId` reaches it.
 
 Controller handler name / OpenAPI operationId: **`revokeSharedAccess`** → generated client method `revokeSharedAccess()`.
 
@@ -114,7 +114,7 @@ Observability: no new metrics. The endpoint is covered by the existing global `M
 
 `GetShareRecipientsDto` SHALL validate `itemId` with the same allowlist as `RevokeSharedAccessDto` (`applications|toolsets|conversations|skills|prompts` prefix, `IsValidFilePath`, `@MaxLength(2048)`) and carries no `resourceKind` field: a resource revoke cannot act on has no count worth answering, and every accepted resource type — prompts included — is now identified by one self-sufficient `itemId`.
 
-`ShareService.getRecipientsCount` SHALL call `getSharedResources({ resourceTypes: [kind(itemId)], with: 'others', includeUserInfo: true })` and pick the one resource out of that set with `countRecipientsByUrl` + `resolveRecipientsCount(counts, itemId, decode(itemId))` — both encodings are tried because list ids and DIAL Core share urls differ in percent-encoding for some resource types. DIAL Core has no single-resource variant of this query. `kind(itemId)` SHALL resolve `skills/`-prefixed ids to `SKILL` and `prompts/`-prefixed ids to `PROMPT` via the existing `RESOURCE_KIND_BY_PREFIX` map, unchanged by this capability.
+`ShareManagementService.getRecipientsCount` SHALL call `getSharedResources({ resourceTypes: [kind(itemId)], with: 'others', includeUserInfo: true })` and pick the one resource out of that set with `countRecipientsByUrl` + `resolveRecipientsCount(counts, itemId, decode(itemId))` — both encodings are tried because list ids and DIAL Core share urls differ in percent-encoding for some resource types. DIAL Core has no single-resource variant of this query. `kind(itemId)` SHALL resolve `skills/`-prefixed ids to `SKILL` and `prompts/`-prefixed ids to `PROMPT` via the existing `RESOURCE_KIND_BY_PREFIX` map, unchanged by this capability.
 
 DIAL Core omits resources nobody currently holds from a **successful** response, so a resource missing from a successful result SHALL answer `0`. An upstream failure SHALL surface as `502`/`503` rather than as a count, leaving the caller to decide how to degrade — a fabricated `0` would silently remove the owner's only way to revoke.
